.travis-requirements.txt
six>=1.4.1
requests>=2.3.0
responses>=0.2.2
python-dateutil>=2.2
xmltodict>=0.9.0
django-nose>=1.2
coverage>=3.7.1